Churn away further at the example AST
Some realisation: the association key-value thing, it needs to be an
expression. If it's not the AST gets way too complicated. But that
should probably mean that it can have a type and such? Maybe just as a
tuple? Or maybe Assoc is the only thing that exists, the 2-tuple is kind
of special, and we don't have other tuples. (What about the 0-tuple,
unit? Well for json compatibility we already need null, so might use
that?)