Implement comprehensions in the evaluator
This does lead to some changes to the AST as well. I think I don't want
to have pairs as first-class values in the runtime, at least not as a
hack to implement dict comprehensions. (Should I call it dict instead of
map, because map is confusing with the function "map"?) So one quite
elegant way in the AST at least, is to give the Seq type two variants,
one for single elements and one for key-value pairs. But a consequence
of that is that my primary example can no longer be expressed in the
AST, because it has let bindings in the middle and those are
expressions. Well, technically they would be let-bindings for the
expression left of the : then, so they would be available to the key but
not to the value. But that is both weird and inconvenient. The solution
then, is to add let bindings as one more special construct to the
comprehension. On the one hand it's a shame to duplicate, but on the
other hand, everything falls into place nicely now. This is basically
what LINQ was in C# too, and if it starts resembling that, that is
usually a good sign.