Work on parsing operators
I need to think this through a bit more. Probably it needs to define the
expression layers so they can be parsed with the right precedence, but
I'd like to take some inspiration from Pony here and disallow most
ambiguous combinations of precedence. We can actually do this with the
CST parser, because it can tell us whether one side of the operator is
enclosed in parens or not -- something that the AST would not be able to
do. But that's for another day.