Add a Bison grammar
To sort out the precedence rules for myself, it's useful to write down
a machine-checked Grammar, where Bison will just point out all my
ambiguities to me.

This did clarify one case for me too: the ambiguity with the let and for
inside a comprehension. The solution is simple, to not take a top-level
expression there, but something a bit more specific that already
excludes the if and the let.