Add pretty reporting of errors
Most of the pretty printing code is adapted from Squiller, which I wrote
and own the copyright to. It is licensed Apache 2.0 as well.

This now makes it much easier to debug the parser because I can *see*
where it runs into an error. Right now it's something related to field
access, maybe I forgot to consume the token.