Implement more abstractions, make string val rc
If we make Ident an Rc<str>, and sometimes idents spill into
Value::String at runtime, then string values have to be Rc<str> as
well. This is unfortunate on the one hand, because now they are
doubly-refcounted in some places (the inner string, and the outer
value). But I hope it's not too bad, we can do performance later.