Begin working on example to invert mapping
I need some kind of group by for this, and for that I need what in
Haskell is called unionWith. I could add that as a builtin on maps,
then I need to implement lambdas to make progress.

But now there is a dilemma. All of that is much nicer syntax-wise and
typecheck-wise if there were freestanding functions instead of methods.
And when you define helpers yourself, they are going to be free-standing
anyway. So maybe it's better to make them free-standing from the start?

But anyway that's for later.