Enable dash as part of identifiers
I want to write a GitHub Actions config in RCL, but they use kebap-case,
and that gets really ugly if I have to quote them, so let's just allow
dashes in identifiers. Why not? Well because x-y is now an identifier
and not the arithmetic expression x - y. But you should put spaces
around your operator anyway so I am fine with this.