Change Span repr to accomodate 262 TiB documents
The 4 GiB limit made me worried, you might have json input files larger
than that. For individual tokens it was still okay if the start offset
is usize but then the len is <=4GiB, if you have a string literal or
something that long, you can break it up. But now that I started
tracking also the enclosing spans, 4 GiB might not cut it. Change the
representation so we can devote 48 bits to the offset, and still 32 to
the document id. So you should be able to load something like Nixpkgs
and have enough document ids and enough span length.