Replace FixmeError for json serialization
Error reporting for json serialization is tricky, we can't necessarily
point to a source location where the offending value was produced. For
example, we may have loaded a string from a file and parsed it ... then
what line do we point to?

To make things nicer, it would be good to have some "path" in the json
structure where we are serializing, to point out where things are going
wrong. But for now, just pointing out that things are wrong *somewhere*
is good enough.