Change Unicode escape from \u{} to \u[]
This is to keep the lexer simple, because {} inside an f-string already
means a hole. We could make an exception so it's not a hole after \u,
but we can also just change the delimiter.

I wonder though, if this is the right call. Rust already uses {}. Best
to not spend the novelty tokens on something trivial like this. Maybe I
should complicate the lexer ...