Begin working on pprint implementation
I was struggling for some time to find good resources on how to
implement a pretty-printer, and then today I discovered Philip Wadler's
paper which has this beautiful approach. So I'm going to implement that
now. The nice thing is too, I could use this both for the autoformatter,
and for printing values and types later for in error messages.