Use pprint for fmt module
This was a very big change, and I had to convert everything at once to
be able to compile, but I am quite pleased so far. It seems to actually
format the examples decently now, aside from putting trailing commas in
collections which I understand why it happens but maybe I should have an
exception for.