Correct outdated doc comment
I now have proper line breaking! This paper by Philip Wadler is amazing;
I expected that a nice composable abstraction for documents would be
possible, but that pretty-printing would be a lot harder (it is a kind
of graph search problem after all). But it turns out the two can go hand
in hand, and the implementation was not even that difficult. So thanks
Philip :-)