Unify the different quote types somewhat
This took a few iterations to make something where it actually factors
nicely, and it's not special cases all the way. Moving the quote style
into the token helped a lot with this in the end. Now Token is no longer
a simple enum, but the code is nicer. I hope rustc is able to inline the
definition and flatten the entire thing, but not sure if it implements
that optimization.