Emit comments as zero-width for formatter
It would matter if comments could be present in wide mode, but actually,
comments already force tall mode so this shouldn't affect the behavior.
Still I think it's nice to be explicit about it. But maybe I should just
remove the variant and remove the complication?

Are there other things that can be zero-width? Hmm ...