Solve the fmt-hanging-if in yet another way
I think that not only do we need the newline to be conditional, also the
indent needs to be conditional on having emitted the newline. This does
feel slightly ad-hoc, but it does enable me to format the ifs the way I
want to.

Also add some goldens to increase the coverage of the pprint module.