Truncate long lines when printing errors
I tried to load a big json document, and it was impossible to make sense
of the error because it printed a line of several megabytes (that
wrapped in my terminal) and then a line of several megabytes of spaces
for indenting the squiggly underline.

We can improve the user experience here by truncating the long line,
so you still have something to make sense of.