Add null
I'm very sorry, but I have to for json compatibility. I want RCL to be
able to be able to generate config for other languages, so null will
have to be there. Fortunately it is not null itself which is so bad, the
real bad thing is when everything is implicitly nullable, and I'm not
going to do that in RCL, so this should be okay.