Add flush break directive to formatter
For a long time I was wondering how to best approach this, but now this
seems like an elegant solution.

I thought I committed this separately, but it was stuck together with an
unrelated change, potentially due to a rebase? I disentangled them after
the fact in this commit now, but the original commit message may be
lost, if I wrote one at all.