Strip indentation from triple quoted strings
This is a squash of five commits that I worked on over the past two
weeks. All of it is a bit of a mess frankly. I haven't discovered an
elegant way to factor out the different string literal behaviors. The
way it is right now has some repetition, but at least with the callbacks
it is not too bad at the call site.

The changes end up being relatively modest in the 'string' module, but
the way to getting here was wild!

Formatting of """-quoted strings is not yet implemented properly.