Escape trailing newlines in formatter
There is a bug/feature/hack in the pretty-printer that makes it eat
trailing whitespaece. This is usually good in expressions, but it is bad
inside string literals because it changes their meaning.

The fuzzer finally caught this, because it breaks idempotency of the
formatter.

Rather than removing the hack and dealing with whitespace properly, I
have another hack. But it's arguably a good hack! Replace the trailing
whitespace in the string with an escape sequence. Those don't get eaten,
so it restores idempotency. But it's also good for humans, because it
makes it obvious and visible that there is something there. Some editors
do highlight this, but not everybody has this set up, so it is also kind
of nice to draw more attention to. On the other hand, maybe we should
not pretend to know what is best for the user, and it's not so obvious
that U+0020 is a space from the source code. We can reconsider later,
but for now I like this solution.