Force a line break after """
Allowing this creates so many problems that I now have to work around
and fix ... let's just ban it, it will be easier.

Supporting """-quoted strings that are sometimes immediately followed by
a line break and sometimes not, and thinking about what that meant for
how to strip the whitespace ... it was a big ugly mess. I almost got it
to work, until it got to the case of

    let x = """   x
        bar
        """;

What should that even mean? So to simplify things, let's just ban this
and demand a newline after the opening quotes. I'm open to later
relaxing to allow single-line """-strings, but for now I just don't want
to implement all the special cases.