Broaden what "blank line" means for indent stripping
Previously only a line with no content would be considered a blank line.
Now we also consider lines with only spaces.

This resolves a nasty non-idempotency in the formatter that was
discovered by the fuzzer. Under the previous definition, the following
string (where I marked spaces with a period):

    """
    .
    .."""

consists of a line with one space and a line with two spaces. The common
indent is therefore one space, which gets stripped, so the contents of
the string is "\n." (again, using dot to denote space). Now we emit
those lines:

    """

    ."""

But this string has a different meaning! Because now the first line
became blank, so the common indent is now one space, and the string
turns into "\n\n". It is a similar case to the one with trailing space,
which I fixed by emitting that space using an escape code. We could do
the same here. But more logical is to consider the first line in the
first case a blank line. Then the first string means "\n\n", and the
formatted version still means that.