Add test for counting common indent
After reviewing my own code before I'm merging it, I thought there would
be a bug in finding the common indent, when *all* the lines are blank.
Then we'd skip over all of them and not update the minimum.

But a blank line has to end in a newline, and there is an indent after
the newline. Or maybe there is no indent after the newline but the end
of the string literal, but that also determines the indent then. So we
are fine.