Add a loader module and query subcommand
My end-goal for today was to make a "query" subcommand, that could act
as a kind of jq alternative. Last week I had to analyze some json
document, and I was struggling to figure out the json incantation to
make it work, but in RCL I could do it in a few lines.

But in order to do that, I had to paste my big json into an RCL
document, bind it to a variable, and then write my expression. It would
have been nice if I could have loaded it from stdin, and then put the
expression on the command line.

So in order to have that, I now really need to load two documents at the
same time. Fortunately I prepared for this with the DocId. But we still
need a way to collect all these documents. This is where the Loader
comes in. You load a document into it, and then you can parse and
evaluate parts. For now it's all done on demand, in the future, when I
add support for imports, it would be good to cache the results.

This also leads to a nice simplification/deduplication of the main
binary.