Fuzz the json output, and idempotency
The json output, it's nice to fuzz it but I kind of trust it's okay, we
have plenty of golden tests and it's not so complex. Still, it's good to
fuzz.

But idempotency, that is pretty neat. It is quite a high-level property
that it verifies: RCL is a superset of json, and a json document should
evaluate to itself. So I expect this fuzzer is going to be valuable.

The property is currently violated though. For starters, there are some
issues around integer literals that I swept under the rug for dealing
with later ... I guess the time has come to deal with them.

The other thing that I expect might bite me is preverving the order in
lists and dicts, although the current order is sorted, so actually that
should work out.