Simplify newline in formatter, use in fuzzer
At first I thought the fuzzer was failing because of the trailing
newline. Turned out that that was not the problem; I was feeding the
full untruncated input into the fuzzer, including the final byte! So a
lot of coverage was missing, the fuzzer never would be able to fuzz
anything that ends properly ... oops! I should have made a coverage
report of the fuzz corpus perhaps.

Now that that is fixed, there were some issues slicing off the last byte
when it's not a code point boundary etc ... so instead let's take the
input as bytes. The downside of this is that the fuzzer will not print
anything pretty when it fails, but that's acceptable, we can just open
the file to see what's in there.