Remove unused print method from pprint::Doc
I thought I'd keep it around for if I want to pretty-print stuff to put
inside error messages and such. But maybe a much nicer idea is to make
the error printer output Doc types to format.

The reason for that is that I can put the color output in the formatter,
and then turning color on or off for NO_COLOR support will also work on
error reporting.