Update the fuzzer to put metadata in a comment
This makes the fuzz input valid files themselves, and it makes them
human-readable as well. This should help a bit to debug when the fuzzer
finds an input.

Previously I wanted to use the shebang line for this, but actually a
comment works in the same way. They are treated exactly the same as
NonCode elements, so there is no reason to prefer one over the other.

The other change is that this now enables the fuzzer to also fuzz
different widths for the formatter.

The downside is that my existing corpus is now useless. I'll write a
small program to migrate it.