Add a few more entries to the fuzz dictionary
I migrated the fuzz corpus, and introduced an integer overflow bug on
purpose. But the fuzzer didn't find it at first, it seems it's not so
effective. So I added the literal for i64::MAX to the dictionary, and
then a minute or so later the fuzzer did crash on the intentional bug.