Color builtins in Vim syntax
I was thinking to syntactically distinguish methods from field access,
but maybe it is not needed if we also have indexing. E.g. when you have

    let x = {
      contains = name => name == "element";
      name = "example";
    };
    x.contains("name")

should this return true or false? I think it's okay if it returns true,
and because to call the inner function, you could write

    x["contains"]("name")

but there is no alternative to reference the method unless we add
syntax for it, like

    x->contains("name")

But I feel this distinction between -> and . will make things ugly, and
possibly confuse newcomers (even though admittedly, reusing the same
syntax for two different constructs is far more confusing if you
actually understand that there are two different things going on).

But anyway, to get back on topic, if the builtin gets a different color
when you type it, that should also help users to understand that
something special is going on.