Fix how a {}-comprehension's type is resolved
We have to look at the syntactic construct, we cannot look at the
runtime value. Otherwise, because an empty collection is a dict, the
following would be a dict:

    {for x in xs: x}

Even though we can tell that it's a set, because it contains a scalar
value in the comprehension, not a key-value pair.