Add Python venv shell hook
So I can get PyCharm to work with this environment. I don't see Pygments
show up in its package list though. But let's see ...

Edit: Okay that did not work at first, but manually creating the
virtualenv and symlinking the site packages into it does work. Ln does
print a warning but for now it's okay.

Edit 2: Nope that was the wrong approach too. The Python binary on its
own is sufficient, and Pip does need to be in there for this to work
with PyCharm. Thanks Laurens for showing me how to do this.