Support strings in Pygments lexer
Only regular strings, not yet f-strings. But this does give me an idea,
I like this unified approach of the Pygments lexer where we also lex
inside the string.

Currently in the Rust lexer we don't care so much what happens inside
string literals, but then later when we interpret them, we have to do
work to unescape, which is secretly kind of a lexer in its own right.

Additionally, the handling of string formatting currently with the holes
is still somewhat ugly. And all that business with the multi-line
strings ... if we could break this down in the lexer (maybe the lexer
should just report the lines), and let the parser put the strings back,
maybe that would be nicer? I guess I'd have to try to see.

One nice thing that could come from it is that the highlighter would
more easily be able to highlight escape sequences inside string
literals. Currently it doesn't, but that will make it more difficult to
verify that the Python and Rust lexers produce the same output.