Remove top-level Pygments from flake
I put it there to test building it, it was not supposed to be included.
Now it gets built by Garnix, but it doesn't need to. I think it will
still get built anyway because the typecheck now depends on it, but it
doesn't need to produce a top-level entry in the CI checks.