Split innermost ast::Seq out into a yield type
This avoids having this one unreachable case. Make invalid states
unrepresentable. I am a bit worried about the size of the enum though,
I'm not sure Rust can fold the two enums back into one and avoid the
additional discriminant. But on the other hand, Seq::For was already
larger, so even if we add a full pointer-sized element, it still does
not make Seq itself larger.