Add support for statements in the parser
This is a squash of many intermediate steps, and a somewhat big commit.
I usually don't like to make commits this big, but this is one of those
changes where you really have to touch everything before the program
even compiles again, and I want everything to compile.

Observation: Both inside Seq and inside Expr, we had "let". Now I'm
adding "assert" and "trace", and they should be allowed in the same
places where let is, because it is a similar construct. To avoid
duplicating everyting, I wanted to extract things into a type. My first
attempt was wrong, but I got pretty far: the parser was done, the CST
was done, the pretty-printer was done. And then when I got to the
abstractor and the AST, I discovered a much more elegant way.

Option 1: Stmt<T>, and use it as Stmt<Expr> and Stmt<Seq>.
Option 2: Stmt and (Stmt, Expr), (Stmt, Seq).

With option 1, I had to make all functions that processed the Stmt
generic, and make them take a function pointer to the function that
would do the inner processing. It was unwieldy everywhere.

After implementing option 1 all the way in the CST, when I get to the
AST I stumble upon option 2. Option 2 is clearly superior, because I was
planning to add let-chains to the CST anyway to be able to format them
as all-tall or all-wide. And with option 2, then we can make the nodes
just a (Vec<Stmt>, T), while with option 1, it's not so obvious.

After going through all of this I am quite pleased with the result. It
now enabled reusing some code between Expr and Seq, and even though the
result is more code, it will help with implementing assert and trace.

This now pipes these constructs through everywhere, but they are not yet
implemented in the evaluator.

As a side effect, there is more grouping in the formatter which causes
one of the golden tests to be wide instead of unfolding tall in its
entirety. I'll fix this later.