Improve formatting of multiple comprehensions
When they each fit on a line, there is no need to make the collection
itself be tall. Though now I'm thinking of it, this could lead to a
situation where even a single comprehension is formatted with Egyptian
brackets. I guess that's okay. So now

    let x = [... really long comprehension...];

could turn into

    let x = [
      ...really long comprehension...
    ];

and that one would still be on a single line. But at least this is only
the case for comprehensions, not for scalar elements. So we basically
treat the comprehension as one scalar element.