Simplify formatting of assert expressions
I realized we don't need the flush indent, which is a subtle thing.
An assertion is just like a function call, but without the parens:

    assert a, b;

formats in tall mode to

    assert
      a,
      b;

But if we had

    assert(a, b);

Then it would format to

    assert(
      a,
      b,
    );

So pretty much the same structure, no hanging indents here or anything
to flush. The flush was silly because we know the line was not empty.