Use box drawing chars for errors, fix column off-by-one
Inspired by Terraform's error output, the box drawing characters tend
to get connected in a terminal (but not IntelliJ it seems), it looks
slightly cleaner than pipes. And these code points are in Unicode since
1993, so I don't think they are particularly controversial to use
(unlike color emoji).

And now that I have to rewrite all the goldens, let's fix the column
counter too. It still counts bytes and not columns, but I'm not sure
if the way it's counted is consistent across applications. We could
the existing unicode-width dependency to count though.