Define AST and CST nodes for import statement
When I first started working on this, I had the CST node contain a
string literal. This ensures that you can't do any too crazy dynamic
imports, and it has some properties that may be useful later:

 * You can statically determine the import graph from just the syntax
   trees, without having to do any evaluation.
 * You can grep for a file in a codebase and find all the places where
   it's imported.

But now I'm rebasing this, and the definition of string literals
changed, so now this is a bit annoying to enforce in the CST/AST. We
could still check for it during evaluation, but I'm not doing that for
now. If things need to become stricter, it'll have to be a breaking
change later on then.