Add regression test for string escaping issue
This is an issue found by the fuzzer. I am somewhat surprised that it
found it only now, after I was fuzzing for several days (though not
intensely) and this bug has been present since the beginning.

Perhaps it was not caught earlier because normally it causes just a
parse error. It required a second string to be present to compensate,
and then it was caught by a non-idempotency in the evaluator.

I think the proper way to fix this is to merge the string lexer into
the original lexer, move all the states there. Maybe they should be
duplicated and in the end yield the token, but maybe they should just
directly yield parts of the string as tokens. I'm interested in that,
because it would enable highlighting escape sequences in the highlight
command, and that would enable verifying it against the Pygments lexer.