Rewrite the lexer to handle strings
This makes the lexer break down string literals and f-string into the
relevant parts. But this does mean it's up to the parser to put them
back together.

This also changes the syntax for enclosed Unicode escape sequences from
\u[xxxx] to \u{xxxx} to align with Rust. Spend the innovation tokens
elsewhere. It turned out for the Pygments lexer it was not really harder
to write things this way, and for the Vim syntax, I think it can be
fixed. Python doesn't have these but it has \N{xxxx} which also use {}
rather than []. So \u{xxxx} it is.