Make lexer error reporting a bit more ergonomic
I finally discovered a nice and composable way to handle the errors with
optional notes! Convert a span into an error type, method chaining to
agument the error if needed, and a final .err() to convert it to a
result. I should do this in the parser as well, but one thing at a time.