Parse string literals from tokens, simplify types
I had already managed to unify the double and triple-quoted strings, and
format- and non-format-strings to a large extent, but the latter two
were still two distinct cases. So what can we do about that? It turns
out that we can unify them pretty much all the way. It does mean that
all string literals turn into Expr::Format in the AST, but I think
that's okay, it does make things easier. It does create extra
indirection and it's likely a performance pessimization, but right now
I am not optimizing for performance.