Add test case for formatter bug
This bug was discovered by the fuzzer -- I assumed that I was breaking
up strings into lines, or at escape sequences, but strings can be broken
at other places too. Not sure where to best fix this, in the lexer or
whether to join them again in the parser. But it's late so this is
enough for today.