Assert that lexer output is on char boundaries
At first I thought to put it in the fuzzer as a check afterwards, but
maybe I forget to run the fuzzer, so let's put it as a debug assertion
in the lexer. Then it will certainly catch when the lexer has a bug (and
maybe that is caught by the fuzzer still).