Add new command-line parser modules
This defines the modules, but does not yet use them. The args iterator
and structure of the parser is taken from Squiller, also written by me,
which in turn takes it from Tako, also written by me. This application
has the most complex command line so far though, so the parser does get
somewhat unwieldy. Perhaps there is a way to simplify it further.