Add alternative macro-based CLI parser
Parsing the CLI into structs into a satisfying way is a lot of work. I
was really on the verge of giving in and pulling in the Argh crate for
this. But then it pulls in syn and proc and that entire circus, which
does make sense for most cases, but I don't quite want to depend
something that heavy yet. So let's see how far we can get using just
macro_rules!. I started writing parsers by hand but it did get unwieldy
quickly, so macro it is.