Pin Nixpkgs to a version with Rust 1.70
I started using the is_terminal method which was introduced in 1.70, but
the flake (and therefore CI) were using an older Rust version that did
not have it yet. It's better for them to be in sync, even if it means
they are not in sync across all of my repositories.

Thanks to nixhub.io for making it easy to find the right Nixpkgs
revision.