Include unit test coverage in coverage report
Previously this only included the golden tests, but it would make sense
to also include unit tests, because for the CLI parser that is a more
appropriate type of test.

Juggling the profraw files is a hassle though. I got it working with
grcov now and it generates a report where the line coverage looks
correct, but function coverage is all in the red now! It looks like
there are now 3× as many total functions listed, but the same number is
covered. The number of binaries went from 1 to 3 so I suspect it's
related to that ... can I make grcov not double-count them?