Add tests for new sandbox CLI parsers, add check
I could use an ADT to make the invalid case of having includes without
sandbox=pure unrepresentable, but on the other hand, storing includes
when they are not used is not so harmful, and it simplifies the parser
to not have to unpack the sandbox mode all the time. So for now I'll add
a check in the parser but I don't move the includes into SandboxMode.