Use comma as the separator in record notation
I am still ambivalent about this. On the one hand I have a strong sense
that "key = value;" is a statement that needs a terminator. On the other
hand, it makes things more uniform to have only a single separator.

I think I just need to get used to the comma, and then I will not mind
so much. But let's see.