Restructure builtin methods, generate boilerplate
This changes the type definition for builtins to no longer require a
Rust closure, but to use a function pointer instead. The cost is that
the method receiver is now part of the value.

Also, we can now generate some of the boilerplate for these method
calls. The error messages are still far from the type errors I would
like to have, and maybe even a regression in some sense, but at least it
needs a lot less code.