Add skeleton for std.load_utf8
It can be called now, but the implementation panics. It was not so easy
to make the fuzzer discover it even with the dictionary. With the
combined "std.load_utf8" entry it does find it quickly, but with
separate "std" and "load_utf8", it might have found it eventually, but
not within a few minutes of fuzzing.