Bring back checking of call arguments
I changed the way calls work, so error messages can work with the spans
that they originate from and report better errors, but this is a
somewhat deep restructuring that made pattern matching on the args less
viable. As a middle ground, add some helpers to check the arity. In most
cases I don't need to check the type, and in the cases where I do, for
now I can do it manually.