Make Clippy happy
I don't think this is really an improvement, but I don't have a very
strong opinion on it and silencing Clippy is as ugly, so let's just do
it.

Also, where I previously had to silence Clippy (for the same lint), I
did end up adderssing that, and the silences are no longer needed.