Move expr_import inwards in the grammar
Previously, with expr_import at the same level as statement-like
expressions, it was not part of expr_op, and inside a sequence, only
expr_op is used, which meant that you couldn't put imports in a
sequence, at least not without parentheses. By pushing expr_import more
inwards to be part of expr_op, it is now allowed inside sequences.

Also add a test for this.