Define lambdas in the grammar
This was a bit of an adventure, to get the grammar right. See also a
discussion of the ideas in ideas/lambdas.rcl added in this commit. For
now I will plow through with the syntax that I like best, let's see if
it works out, we can always change it later.