"Solve" the Bison reduce/reduce conflict
I think I understand where it comes from now: Bison only has a single
lookahead token, and when it encounters the ) it doesn't have enough
lookahead to know how to reduce.

The "fix" is to disallow lambdas of the form `(x) => x` in the Bison
grammar, and require them to be written as `(x,) => x`. It's a bit ugly,
but it works for now. At least I understand the conflict, and I think
it's mostly harmless for the parser in practice.

I don't know if this will become problematic in other places. I can
imagine that a Tree Sitter grammar may become difficult because of this,
or something. But let's worry about that later.