Implement List.group_by method
I implemented it ... and it worked! I'm kind of surprised. I haven't
fully tested the edge cases yet though. This is a big step forward in
RCL ergonomics, no more need to write the group_by by hand. Next up will
be key_by for when the key has to be unique.