Ensure import execute in a clean environment
And also, stop re-creating the stdlib instance for every expression that
gets evaluated. That should lead to a small performance improvement as
well; profiling my very inefficient Advent of Code solution,
stdlib::initialize showed up on the profile. Not extremely hot, but a
few percent of execution time, and completely pointless.