Put a maximum length on std.range size
We materialize the entire list, so putting in a very large value is a
footgun to cause OOM.

There are other ways to cause OOM, in particular by making a cartesian
product in a list comprehension. Maybe I should limit the list size in
comprehensions too, possibly to the same limit. But that's for later,
for now bound the list in size.

This issue was discovered by the fuzzer.