Allow omitting the input file from the CLI
I want to use RCL as a jq alternative, and one thing that still gets in
the way slightly is to have to type - as input all the time. That is
fixable by making the file optional and defaulting to stdin. It even
works for 'rcl query'.

A downside of this, is that new users might try to run 'rcl evaluate',
and it will sit there waiting for an EOF, but the user doesn't know that
and think it hangs.

_That_ problem can be solved by only defaulting to stdin if stdin is not
a TTY. But this now creates a weird context-dependent behavior and makes
the application unpredictable. We already have that with colors though.
I'm leaning towards that this is fine; if you want fully predictable
behavior you can still get that by specifying - as the input file.