Consider adjusting the if-then-else syntax
After working with the current form for a bit, I think I like the block
form better because it's easier to format, even though on a single line
it may be a bit uglier. It is also more consistent with the if syntax
inside comprehensions. And Nim uses this syntax, so that gives it some
legitimacy.