Simplify stdin defaults
I just had an idea: instead of providing the state of stdin as an input
to the CLI parser, simplify the CLI parser (and tests) enormously by
removing that again, but instead, when it picks stdin as the file,
distinguish between "stdin (explictly)" and "stdin (default)".

Then in the input handling, I had another idea to make things simpler
without being a complete footgun to newcomers: default to stdin always,
regardless of whether stdin is a TTY. But, when stdin was chosen by
default and it is a TTY, print an informational message to stderr to
point out to users that we're waiting for input from stdin.