Add test case for stack overflow
I discovered this while trying to implement and Advent of Code problem
in RCL. Everything appears to work, except that it overflows its stack.
This is tough to solve, but at least we can make it error niceley
instead of crashing. I am a bit disappointed that the fuzzer has not
discovered this yet.