Add an --output=raw output format
I've needed this a few times, when I want to print a table from json
values. It's not too difficult to add. This implementation is somewhat
inefficient because it constructs and then formats an intermediate Doc
while it could print straight to stdout, but this does make the output
modes more uniform, and it makes it slightly easier to report errors
about non-formattable values. Also, the overhead is no worse than json
output, which so far has been fine for me if the document is not
humongous.