Add String.chars builtin method
Until now you could fake it by accident using str.split(""), but I'm not
really happy about that behavior, and it creates an empty string at the
start and the end. It was useful for solving the Advent of Code though,
so let's add this first-class.

Now that we have this, probably I should also change String.len to
return the number of code points and not the length of the UTF-8
encoding. And then create a fuzzer that verifies that the following
diagram commutes:

    String   ------- len -->   Int
       |                        |
     chars                      id
       |                        |
       v                        v
    List[String]  -- len -->   Int

Also for indexing and slicing into strings, we shoudl slice by char
then, so you don't have the problems like in Rust that you can slice a
code point and you need to handle that error case. It does come at a
performance cost, but for a configuration language that is fine.