Make String.len count code points, not bytes
This ensures that

    foreach s: String: s.len() == s.chars().len()

which seems like a desirable property to me. In a low-level language
like Rust, len() counting the UTF-8 byte length makes sense, but in a
high-level configuration language, it makes more sense to me to count
code points. One could argue that if we go down this slippery slope,
it should count grapheme clusters instead, but the definition of those
is not stable over time and depends on geopolitical considerations of
$CURRENT_YEAR, I don't want to go there.