Change if-then-else syntax to use : instead of then
After having used RCL in practice for a few months, I think I'm leaning
towards this. It resolves an awkward way of formatting the if-then-else
multi-line, and it's more consistent with 'if' inside a comprehension.
It also makes the syntax resemble Python more for the multi-line case.

What was holding me back previously is that I think the colons on a
single line look kind of awkward, e.g.

    let x = if cond: true-val else: false-val;

But I expect I will get used to it. What pushed me over the edge was
that Nim uses this syntax. Though Nim is maybe not the best
justification to cite here, from my very limited experience using it,
it looks like a kitchen sink of syntax, which is a valid point in the
design space (Perl and Raku have fans too after all) but the opposite
of what I want RCL to be.

But I do think this change brings more consistency and regularity. RCL
is more colon/delimiter based, and less keyword-based. No 'begin end'
but '[]', and no 'then' but ':' fits with that.